backend/requirements.txt
flask==3.0.2
werkzeug==3.0.1
python-dotenv==1.0.1
openai==1.12.0
pytesseract==0.3.10
Pillow==10.2.0
PyPDF2==3.0.1
flask-cors==4.0.0 